+++ src/main.cpp
@@ -3,7 +3,7 @@
 
 LD::LightData ld(10/*Recieve Pin*/,11/*transmit Pin*/, 1250/*transmit speed*/);
 LD::LDStream* LD::LDStream::instance = &ld; //this sets the instance pointer to the correct instance of the class (needed for the ISR function)
-int LD::Session_var::id_counter=0;
+short int LD::Session_var::id_counter=0;
 
 
 void setup() {
